INTYGFV-15155: Added QuestionUnderlag to InternalToCertificate & CertificateToInternal. Implemented AbstractQuestionUnderlag.
