+++ fk/luse/src/main/java/se/inera/intyg/common/luse/v1/model/converter/InternalToCertificate.java
@@ -25,6 +25,7 @@ import se.inera.intyg.common.luse.v1.model.converter.certificate.category.Catego
 import se.inera.intyg.common.luse.v1.model.converter.certificate.question.QuestionAnnatBeskrivning;
 import se.inera.intyg.common.luse.v1.model.converter.certificate.question.QuestionKannedomOmPatient;
 import se.inera.intyg.common.luse.v1.model.converter.certificate.question.QuestionMotiveringTillInteBaseratPaUndersokning;
+import se.inera.intyg.common.luse.v1.model.converter.certificate.question.QuestionUnderlag;
 import se.inera.intyg.common.luse.v1.model.converter.certificate.question.QuestionUtlatandeBaseratPa;
 import se.inera.intyg.common.luse.v1.model.internal.LuseUtlatandeV1;
 import se.inera.intyg.common.services.texts.CertificateTextProvider;
@@ -54,6 +55,9 @@ public class InternalToCertificate {
             )
             .addElement(
                 QuestionKannedomOmPatient.toCertificate(internalCertificate.getKannedomOmPatient(), index++, textProvider)
+            )
+            .addElement(
+                QuestionUnderlag.toCertificate(internalCertificate.getUnderlag(), index++, textProvider)
             )
             .build();
     }
