+++ fk/luse/src/test/java/se/inera/intyg/common/luse/v1/model/converter/InternalToCertificateTest.java
@@ -28,6 +28,7 @@ import static se.inera.intyg.common.luse.v1.model.converter.RespConstants.GRUNDF
 import static se.inera.intyg.common.luse.v1.model.converter.RespConstants.GRUNDFORMU_CATEGORY_ID;
 import static se.inera.intyg.common.luse.v1.model.converter.RespConstants.KANNEDOM_SVAR_ID_2;
 import static se.inera.intyg.common.luse.v1.model.converter.RespConstants.MOTIVERING_TILL_INTE_BASERAT_PA_UNDERLAG_DELSVAR_ID_1;
+import static se.inera.intyg.common.luse.v1.model.converter.RespConstants.UNDERLAG_SVAR_ID_4;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
@@ -110,5 +111,11 @@ class InternalToCertificateTest {
     void shallIncludeQuestionKannedomOmPatient() {
         final var actualCertificate = internalToCertificate.convert(internalCertificate, textProvider);
         assertEquals(4, actualCertificate.getData().get(KANNEDOM_SVAR_ID_2).getIndex());
+    }
+
+    @Test
+    void shallIncludeQuestionUnderlag() {
+        final var actualCertificate = internalToCertificate.convert(internalCertificate, textProvider);
+        assertEquals(5, actualCertificate.getData().get(UNDERLAG_SVAR_ID_4).getIndex());
     }
 }
